src/backend/requirements.txt
fastapi
uvicorn[standard]
joblib
pandas
numpy
pydantic
scikit-learn==1.6.1
shap
groq
python-dotenv